Add files via upload
Week 1 Upload

Part1) Write a Java program which assigns your name to one variable and some other numerical value you want to share to another variable (e.g. age, number of pets,  number of children, etc.). Then print a greeting along with the value of the name variable to the screen on one line and a message that uses the other variable with numerical value to another line.

For example, my program would be: 

Hello class, my name is: Dr. Mcfadden
I have visited 20 states

Submit your program as an attached  .java file and post a screen shot to show that you have been able to successfully run that program. Make sure you submission adheres to the SubmissionRequirements document.

Be sure to create a program different from any of the programs already posted by your classmates or the examples in class materials.