Week2 Assignment 1
Write a Java program which:
(1) Prompts a user to enter customer id, unit price in this format (e.g. 3.75), quantity (as whole number), product description, and discount in this format (e.g., .10) (use Scanner for input).
(2) Calculates the customer's overall order total before and after discount.
(3) Displays the input data along with the order total before and after the discount to the console.